% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/label_conversion.R
\name{as_labelled}
\alias{as_labelled}
\title{Convert vector to labelled class}
\usage{
as_labelled(x, add.class = FALSE)
}
\arguments{
\item{x}{a variable (vector), \code{data.frame} or \code{list} of variables
that should be converted to \code{\link[haven]{labelled}}-class
objects.}

\item{add.class}{logical, if \code{TRUE}, \code{x} preserves its former
\code{class}-attribute and \code{labelled} is added as additional
attribute. If \code{FALSE} (default), all former \code{class}-attributes
will be removed and the class-attribute  of \code{x} will only
be \code{labelled}.}
}
\value{
\code{x}, as \code{\link[haven]{labelled}}-class object.
}
\description{
Convert vector to labelled class
}
\examples{
data(efc)
str(efc$e42dep)

x <- as_labelled(efc$e42dep)
str(x)
summary(x)

x <- as_labelled(efc$e42dep, add.class = TRUE)
str(x)
summary(x)

a <- c(1, 2, 4)
x <- as_labelled(a, add.class = TRUE)
str(x)
summary(x)
}

