% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{descr}
\alias{descr}
\title{Basic descriptive statistics}
\usage{
descr(.data, ...)
}
\arguments{
\item{.data}{A vector or a data frame.}

\item{...}{One or more variables of \code{.data} that should be printed as table.}
}
\value{
A data frame with basic descriptive statistics, derived from the
        \code{\link[psych]{describe}}-function.
}
\description{
This function wraps the \code{\link[psych]{describe}}-function
             and prints a basic descriptive statistic, including variable labels.
}
\examples{
data(efc)
descr(efc, e17age, c160age)

library(dplyr)
efc \%>\% select(e42dep, e15relat, c172code) \%>\% descr()

}

