% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/efc.R
\docType{data}
\name{efc2}
\alias{efc2}
\title{Sample dataset from the EUROFAMCARE project}
\description{
Sample dataset from the EUROFAMCARE project
}
\examples{
# Attach EFC-data
data(efc2)

\dontrun{
# show variables
view_spss(efc2)

# print frq of dependency
sjt.frq(efc2$e42dep)}
}
\references{
\url{http://www.uke.de/eurofamcare/}
}
\keyword{data}

