% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/efc.R
\docType{data}
\name{efc3}
\alias{efc3}
\title{Sample dataset from the EUROFAMCARE project}
\description{
Sample dataset from the EUROFAMCARE project
}
\examples{
# Attach EFC-data
data(efc3)

str(efc3$e15relat)
table(efc3$e15relat)

\dontrun{
# print frq of relationships
sjt.frq(efc3$e15relat)}
}
\references{
\url{http://www.uke.de/eurofamcare/}
}
\keyword{data}

