% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getter.R
\name{get_na}
\alias{get_na}
\title{Retrieve missing values of labelled variables}
\usage{
get_na(x)
}
\arguments{
\item{x}{a variable (vector) with attached value labels and missing values
(see \code{\link[haven]{labelled}}).}
}
\value{
The missing values associated with value labels from \code{x},
          or \code{NULL} if \code{x} has no missing value attribute.
}
\description{
This function retrieves the value codes associated with missing values
               of variables of an imported SPSS, SAS or STATA data set (via \code{\link{read_spss}},
               \code{\link{read_sas}} or \code{\link{read_stata}}), where missing
               values have not been replaced with \code{NA}s after import.
}
\details{
See 'Details' in \code{\link{get_values}}.
}
\examples{
library(haven)

# create labelled factor, with missing flag
x <- labelled(c("M", "M", "F", "X", "N/A"),
              c(Male = "M", Female = "F",
                Refused = "X", "Not applicable" = "N/A"),
              c(FALSE, FALSE, TRUE, TRUE))
get_na(x)

# create labelled integer, with missing flag
x <- labelled(c(1, 2, 1, 3, 4, 1),
             c(Male = 1, Female = 2, Refused = 3, "N/A" = 4),
             c(FALSE, FALSE, TRUE, TRUE))
get_na(x)
}
\seealso{
\code{\link{get_labels}} to get value labels, or \code{\link{get_values}}
           to get values associated with labels; see \code{\link{set_na}} to
           replace specific values with \code{NA} and \code{\link{to_na}} to
           convert missing value codes into \code{NA}.
}

