% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_even.R
\name{is_even}
\alias{is_even}
\alias{is_odd}
\title{Check whether value is even or odd}
\usage{
is_even(x)

is_odd(x)
}
\arguments{
\item{x}{Numeric vector or single numeric value, or a data frame or list with
such vectors.}
}
\value{
\code{is_even()} returns \code{TRUE} for each even value of \code{x}, \code{FALSE} for
          odd values. \code{is_odd()} returns \code{TRUE} for each odd value of \code{x}
          and \code{FALSE} for even values.
}
\description{
Checks whether \code{x} is an even or odd number. Only
               accepts numeric vectors.
}
\examples{
is_even(4)
is_even(5)
is_even(1:4)

is_odd(4)
is_odd(5)
is_odd(1:4)

}

