% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_float.R
\name{is_float}
\alias{is_float}
\title{Check if a variable is of (non-integer) double type}
\usage{
is_float(x)
}
\arguments{
\item{x}{A value, vector or data frame.}
}
\value{
\code{TRUE} if \code{x} is a floating value (non-integer double),
   \code{FALSE} otherwise.
}
\description{
This function checks whether an input vector or value is a
   numeric non-integer (double), depending on fractional parts of the value(s).
}
\examples{
is.double(4)
is_float(4)

is_float(4.2)

is_float(iris)

}
