% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/checker.R
\name{is_labelled}
\alias{is_labelled}
\title{Check whether object is of class "labelled"}
\usage{
is_labelled(x)
}
\arguments{
\item{x}{an object}
}
\value{
Logical, \code{TRUE} if \code{any(class(x))} is \code{labelled},
          \code{FALSE} otherwise.
}
\description{
This function checks whether \code{x} is of class \code{labelled}.
}

