% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_labelled.R
\name{lbl_df}
\alias{lbl_df}
\title{Create a labelled data frame}
\usage{
lbl_df(x)
}
\arguments{
\item{data}{A data frame.}
}
\value{
\code{data}, with \code{lbl_df} class-attribute.
}
\description{
This method wraps a local data frame and adds a \code{lbl_df} class
               attribute. Printing a \code{lbl_df}-data frame is comparable
               to printing \code{\link[dplyr]{tbl_df}} objects, but the class
               information in the output is replaced by the variable label.
}
\examples{
data(efc)
library(dplyr)
mydf <- lbl_df(efc \%>\%
  select(e15relat, e16sex, e17age) \%>\%
  slice(1:3))

mydf

# or...
mydf <- efc \%>\%
  select(e15relat, e16sex, e17age) \%>\%
  slice(1:3)

lbl_df(mydf)


library(dplyr)
library(sjmisc)
data(diamonds)

mydf <- lbl_df(tbl_df(diamonds) \%>\%
  select(cut, carat, price) \%>\%
  set_label(c("", "Kt", "EUR")) \%>\%
  slice(1:3))

mydf

}

