% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_df.R
\name{merge_df}
\alias{merge_df}
\title{Merge labelled data frames}
\usage{
merge_df(x1, x2, id = NULL)
}
\arguments{
\item{x1}{First data frame to be merged.}

\item{x2}{Second data frame to be merged.}

\item{id}{Optional name for ID column that will be created to indicate the
source data frames for appended rows.}
}
\value{
A full joined data frame.
}
\description{
Merges (full join) two data frames and preserve value and variable labels.
}
\details{
This function merges two data frames, where equal named columns
           will be joined together. Rows of merged data from unique columns
           in \code{x1} will be filled with \code{NA} (and vice versa for
           \code{x2}). This function is a convenient wrapper for
           \code{merge(x1, x2, all = TRUE}, however, unlike base
           \code{\link{merge}}, this function preserves value and
           variable labels. For more details on this function, see
           also 'Details' in \code{\link{merge}} about \code{all}-argument.
}
\examples{
library(dplyr)
data(efc)
x1 <- efc \%>\% select(1:5) \%>\% slice(1:10)
x2 <- efc \%>\% select(3:7) \%>\% slice(1:10)

mydf <- merge_df(x1, x2)
mydf
str(mydf)

\dontrun{
library(sjPlot)
view_df(mydf)}

}

