% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rec_pattern.R
\name{rec_pattern}
\alias{rec_pattern}
\title{Create recode pattern for 'rec' function}
\usage{
rec_pattern(from, to, width = 5, other = NULL)
}
\arguments{
\item{from}{Minimum value that should be recoded.}

\item{to}{Maximum value that should be recoded.}

\item{width}{Numeric, indicating the range of each group.}

\item{other}{String token, indicating how to deal with all other values
that have not been captured by the recode pattern. See 'Details'
on the \code{emph}-token in \code{\link{rec}}.}
}
\value{
A list with two values:
          \itemize{
           \item{pattern} string pattern that can be used as \code{recodes} argument for the \code{\link{rec}}-function.
           \item{labels} the associated values labels that can be used with \code{\link{set_labels}}.
          }
}
\description{
Convenient function to create a recode pattern for the
               \code{\link{rec}} function, which recodes (numeric)
               vectors into smaller groups.
}
\examples{
rp <- rec_pattern(1, 100)
rp

data(efc)
table(efc$c160age, exclude = NULL)
table(rec(efc$c160age, rp$pattern), exclude = NULL)

x <- rec(efc$c160age, rp$pattern)
set_labels(x) <- rp$labels
frq(as_labelled(x))

}

