% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfunctions.R, R/re-exports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{set_na}
\alias{to_character}
\alias{to_label}
\alias{to_numeric}
\alias{to_factor}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}

  \item{sjlabelled}{\code{\link[sjlabelled]{set_na}}, \code{\link[sjlabelled:as_label]{to_character}}, \code{\link[sjlabelled:as_factor]{to_factor}}, \code{\link[sjlabelled:as_label]{to_label}}, \code{\link[sjlabelled:as_numeric]{to_numeric}}}
}}

