% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sjStatistics.R
\name{rmse}
\alias{rmse}
\title{Compute root mean squared error (RMSE)}
\usage{
rmse(fit)
}
\arguments{
\item{fit}{a fitted linear model of class \code{\link{lm}},
\code{\link[lme4]{merMod}} (lme4) or \code{\link[nlme]{lme}} (nlme).}
}
\value{
The root mean squared error of \code{fit}.
}
\description{
Compute root mean squared error  of fitted linear (mixed effects) models.
}
\examples{
data(efc)
fit <- lm(barthtot ~ c160age + c12hour, data=efc)
rmse(fit)
}
\seealso{
\code{\link{cv}}
}

