% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_num.R
\name{round_num}
\alias{round_num}
\title{Round numeric variables in a data frame}
\usage{
round_num(x, digits = 0)
}
\arguments{
\item{x}{A vector or data frame.}

\item{digits}{Numeric, number of decimals to round to.}
}
\value{
\code{x} with all numeric variables rounded.
}
\description{
\code{round_num()} rounds numeric variables in a data frame
   that also contains non-numeric variables. Non-numeric variables are
   ignored.
}
\examples{
data(iris)
round_num(iris)

}
