% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_col.R
\name{seq_col}
\alias{seq_col}
\alias{seq_row}
\title{Sequence generation for column or row counts of data frames}
\usage{
seq_col(x)

seq_row(x)
}
\arguments{
\item{x}{A data frame.}
}
\value{
A numeric sequence from 1 to number of columns or rows.
}
\description{
\code{seq_col(x)} is a convenient wrapper for \code{seq_len(ncol(x))},
   while \code{seq_row(x)} is a convenient wrapper for \code{seq_len(nrow(x))}.
}
\examples{
data(iris)
seq_col(iris)
seq_row(iris)

}
