% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sjStatistics.R
\name{std_e}
\alias{std_e}
\title{Standard Error for variables}
\usage{
std_e(x)
}
\arguments{
\item{x}{a (numeric) vector / variable or a data frame.}
}
\value{
The standard error of variable \code{x}, or for each variable
          if \code{x} is a data frame.
}
\description{
Compute standard error for variable or for all variables
               of a data frame.
}
\examples{
std_e(rnorm(n = 100, mean = 3))

data(efc)
std_e(efc[, 1:3])
}

