% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_contains.R
\name{str_contains}
\alias{str_contains}
\title{Check if string contains pattern}
\usage{
str_contains(x, pattern, ignore.case = FALSE)
}
\arguments{
\item{x}{Character string where matches are sought.}

\item{pattern}{Character string to be matched in \code{x}.}

\item{ignore.case}{Logical, whether matching should be case sensitive or not.}
}
\value{
\code{TRUE} if \code{x} contains \code{pattern}.
}
\description{
This functions checks whether a string \code{x} contains
               the string \code{pattern}. By default, this function is
               case sensitive.
}
\examples{
str_contains("hello", "hel")
str_contains("hello", "hal")

str_contains("hello", "Hel")
str_contains("hello", "Hel", ignore.case = TRUE)


}

