% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlabel.R
\name{unlabel}
\alias{unlabel}
\title{Convert labelled vectors into normal classes}
\usage{
unlabel(x)
}
\arguments{
\item{x}{A data frame, which contains \code{\link[haven]{labelled}} class
vectors or a single vector of class \code{labelled}.}
}
\value{
A data frame or single vector (depending on \code{x}) with common object classes.
}
\description{
This function converts \code{\link[haven]{labelled}} class vectors
               into a generic data format, which means that simply all \code{\link[haven]{labelled}}
               class attributes will be removed, so all vectors / variables will most
               likely become \code{\link{atomic}}. Additionally, \code{tbl_df} and
               \code{tbl} class attributes will be removed from data frames, and
               a \code{\link{lbl_df}} class attribute will be added. See 'Note'.
}
\note{
This function is deprecated. Please use \CRANpkg{sjlabelled} for
      labelled data functions now.
}
