% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_na.R
\name{has_na}
\alias{has_na}
\title{Check if variables have missing / infinite values}
\usage{
has_na(x, ..., out = c("table", "df", "index"))
}
\arguments{
\item{x}{A data frame.}

\item{...}{Optional, unquoted names of variables that should be selected for
further processing. Required, if \code{x} is a data frame (and no
vector) and only selected variables from \code{x} should be processed.
You may also use functions like \code{:} or tidyselect's
\code{\link[tidyselect]{select_helpers}}.
See 'Examples' or \href{../doc/design_philosophy.html}{package-vignette}.}

\item{out}{Output (return) format of the results. May be abbreviated.}
}
\value{
If \code{x} is a vector, returns \code{TRUE} if \code{x} has any missing
  or infinite values. If \code{x} is a data frame, returns \code{TRUE} for
  each variable that has any missing or infinite values. If \code{out = "table"},
  results are returned as data frame, with column number, variable name and
  label, and a logical vector indicating if a variable has missing values or
  not. However, it's printed in colors, with green rows indicating that a
  variable has no missings, while red rows indicate the presence of missings
  or infinite values. If \code{out = "index"}, a named vector is returned.
}
\description{
This functions checks if variables in a data frame have \code{NA},
  \code{NaN} or \code{Inf} values.
}
\examples{
data(efc)
has_na(efc$e42dep)
has_na(efc, e42dep, tot_sc_e, c161sex)
has_na(efc)

has_na(efc, e42dep, tot_sc_e, c161sex, out = "index")
has_na(efc, out = "df")

}
