% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_even.R
\name{is_even}
\alias{is_even}
\title{Check whether value is even}
\usage{
is_even(x)
}
\arguments{
\item{x}{Numeric vector or single numeric value.}
}
\value{
\code{TRUE} for each even value of \code{x}, \code{FALSE} for
          odd values.
}
\description{
Checks whether \code{x} is an even number or not. Only
               accepts numeric vectors.
}
\examples{
is_even(4)
is_even(5)
is_even(1:4)

}
\seealso{
\code{\link{is_odd}}
}

