% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_labelled.R
\name{is_labelled}
\alias{is_labelled}
\title{Check whether object is of class "labelled"}
\usage{
is_labelled(x)
}
\arguments{
\item{x}{An object.}
}
\value{
Logical, \code{TRUE} if \code{x} inherits from class \code{labelled},
          \code{FALSE} otherwise.
}
\description{
This function checks whether \code{x} is of class \code{labelled}.
}
