% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudo_r2.R
\name{r2}
\alias{r2}
\title{Compute R-squared of linear (mixed) models}
\usage{
r2(x)
}
\arguments{
\item{x}{Fitted linear (mixed) model, or a \code{glm} object.}
}
\value{
For linear models, the r-squared and adjusted r-squared value. For
          linear mixed models, the r-squared and Omega-squared value.
          For \code{glm} objects, the \code{\link{pseudo_r2}} is returned.
}
\description{
Compute R-squared values of linear (mixed) models, or pseudo-R-squared
               values for glm-objects.
}
\note{
For linear models, the r-squared and adjusted r-squared value is returned.
        For linear mixed models, an r-squared approximation by computing the
        correlation between the fitted and observed values, as suggested by
        Byrnes (2008), is returned as well as the Omega-squared value as
        suggested by Xu (2003).
}
\examples{
library(lme4)
fit <- lmer(Reaction ~ Days + (Days | Subject), sleepstudy)
r2(fit)

data(efc)
fit <- lm(barthtot ~ c160age + c12hour, data = efc)
r2(fit)

# Pseudo-R-squared values
efc$services <- dicho(efc$tot_sc_e, "v", 0, as.num = TRUE)
fit <- glm(services ~ neg_c_7 + c161sex + e42dep,
           data = efc,
           family = binomial(link = "logit"))
r2(fit)


}
\references{
\itemize{
              \item \href{http://glmm.wikidot.com/faq}{DRAFT r-sig-mixed-models FAQ}
              \item Byrnes, J. 2008. Re: Coefficient of determination (R^2) when using lme(). \href{http://thread.gmane.org/gmane.comp.lang.r.lme4.devel/684}{gmane.comp.lang.r.lme4.devel}
              \item Xu, R. 2003. Measuring explained variation in linear mixed effects models. Statist. Med. 22:3527-3541. \url{doi:10.1002/sim.1572}
            }
}

