% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sjImportSPSS.R
\name{remove_labels}
\alias{remove_labels}
\title{Remove value and variable labels from vector or data frame}
\usage{
remove_labels(x)
}
\arguments{
\item{x}{vector or data.frame with variable and/or value label attributes}
}
\value{
\code{x} with removed value and variable label attributes.
}
\description{
This function removes value and variable label attributes
               from a vector or data.frames. These attributes are typically
               added to variables when importing foreign data (see
               \code{\link{read_spss}}) or manually adding label attributes
               with \code{\link{set_val_labels}}.
}
\examples{
data(efc)
str(efc)
str(remove_labels(efc))
}
\seealso{
\code{\link{add_labels}}
}

