% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_na.R
\name{replace_na}
\alias{replace_na}
\alias{replace_na<-}
\title{Replace NA with specific values}
\usage{
replace_na(x, value, na.label = NULL, tagged.na = NULL)

replace_na(x, na.label = NULL) <- value
}
\arguments{
\item{x}{Variable (vector), \code{data.frame} or \code{list} of variables where
missing values should be replaced with \code{value}.}

\item{value}{Value that will replace the \code{\link{NA}}'s.}

\item{na.label}{Optional character vector, used to label the NA-value (i.e.
adding a \code{labels} attribute to \code{x}).}

\item{tagged.na}{Optional, specifies a \code{\link[haven]{tagged_na}} value
that will be replaced by \code{value}. Herewith it is possible
to replace only specific \code{NA} values of \code{x}.}
}
\value{
\code{x}, where \code{NA}'s are replaced with \code{value}.
}
\description{
This function replaces NA's of a variable, data frame
               or list of variables with \code{value}.
}
\note{
Value and variable label attributes (see, for instance, \code{\link{get_labels}}
        or \code{\link{set_labels}}) are preserved.
}
\examples{
data(efc)
table(efc$e42dep, exclude = NULL)
table(replace_na(efc$e42dep, 99), exclude = NULL)

dummy <- list(efc$c82cop1, efc$c83cop2, efc$c84cop3)
# show original distribution
lapply(dummy, table, exclude = NULL)
# show variables, NA's replaced with 99
lapply(replace_na(dummy, 99), table, exclude = NULL)

library(haven)
x <- labelled(c(1:3, tagged_na("a", "c", "z"), 4:1),
              c("Agreement" = 1, "Disagreement" = 4, "First" = tagged_na("c"),
                "Refused" = tagged_na("a"), "Not home" = tagged_na("z")))
# get current NA values
x
get_na(x)

# replace only the NA, which is tagged as NA(c)
replace_na(x, 2, tagged.na = "c")
get_na(replace_na(x, 2, tagged.na = "c"))

}
\seealso{
\code{\link{set_na}} for setting \code{NA} values, \code{\link{rec}}
           for general recoding of variables and \code{\link{recode_to}}
           for re-shifting value ranges.
}

