\encoding{UTF-8}
\name{sjmisc-package}
\alias{sjmisc-package}
\alias{sjmisc}
\docType{package}
\title{Miscellaneous Data Management Tools}
\description{
This package contains utility functions that are useful when carrying out data analysis or basic statistical tests, performing common recode and data transformation tasks or working with labelled data (especially intended for people coming from SPSS and/or who are new to R).
\cr \cr
Basically, this package covers four domains of functionality:
\itemize{
  \item reading and writing data between other statistical packages (like SPSS) and R, based on the haven and foreign packages
  \item hence, this package also includes functions to make working with labelled data easier
  \item frequently used statistical tests and computation of statistical coefficients, or at least convenient wrappers for such test functions
  \item frequently applied recoding and variable transformation tasks
}
}
\details{
\tabular{rl}{
Package: \tab sjmisc\cr
Type: \tab Package\cr
Version: \tab 1.0.2-10\cr
Date: \tab 2015-06-06\cr
License: \tab GPL-3\cr
}
}
\author{
Daniel Lüdecke \email{d.luedecke@uke.de}
}
\references{
  \itemize{
    \item Documentation: \url{http://www.strengejacke.de/sjPlot}
    \item Github: \url{https://github.com/sjPlot/sjmisc}
    \item Bug reports: \url{https://github.com/sjPlot/sjmisc/issues}
    }
}
