\encoding{UTF-8}
\name{sjmisc-package}
\alias{sjmisc-package}
\alias{sjmisc}
\docType{package}
\title{Miscellaneous Data Management Tools}
\description{
Collection of several utility functions for reading or writing data,
recoding and labelling variables and some frequently used statistical
tests.
}
\details{
\tabular{rl}{
Package: \tab sjmisc\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2015-04-18\cr
License: \tab GPL-3\cr
}
}
\author{
Daniel Lüdecke \email{d.luedecke@uke.de}
}
\references{
  \itemize{
    \item Documentation: \url{http://www.strengejacke.de/sjPlot}
    \item Github: \url{https://github.com/sjPlot/sjmisc}
    \item Bug reports: \url{https://github.com/sjPlot/sjmisc/issues}
    }
}
