\encoding{UTF-8}
\name{sjmisc-package}
\alias{sjmisc-package}
\alias{sjmisc}
\docType{package}
\title{Data Transformation and Labelled Data Utility Functions}
\description{
Collection of miscellaneous utility functions (especially intended for people coming from other statistical software packages like 'SPSS', and/or who are new to R), supporting following common tasks:
\itemize{
  \item Reading and writing data between R and other statistical software packages like 'SPSS', 'SAS' or 'Stata' and working with labelled data; this includes easy ways to get and set label attributes, to convert labelled vectors into factors (and vice versa), or to deal with multiple declared missing values etc.
  \item Data transformation tasks like recoding, dichotomizing or grouping variables, setting and replacing missing values.
  \item Convenient functions to perform frequently used statistical tests, or to calculate various commonly used statistical coefficients.
}
See also package-vignettes.
}
\details{
\tabular{rl}{
Package: \tab sjmisc\cr
Type: \tab Package\cr
Version: \tab 1.4\cr
Date: \tab 2016-01-13\cr
License: \tab GPL-3\cr
}
}
\author{
Daniel Lüdecke \email{d.luedecke@uke.de}
}
\references{
  \itemize{
    \item Documentation: \url{http://www.strengejacke.de/sjPlot}
    \item Github: \url{https://github.com/sjPlot/sjmisc}
    \item Bug reports: \url{https://github.com/sjPlot/sjmisc/issues}
    }
}
