% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/label_conversion.R
\name{to_fac}
\alias{to_fac}
\alias{to_factor}
\title{Convert variable into factor and keep value labels}
\usage{
to_fac(x, drop.na = FALSE)

to_factor(x, drop.na = FALSE)
}
\arguments{
\item{x}{A (numeric or atomic) variable or a data frame with
(numeric or atomic) variables.}

\item{drop.na}{logical, if \code{TRUE}, all types of missing value codes are
converted into NA before \code{x} is converted as factor. If
\code{FALSE}, missing values will be left as their original codes.
See 'Examples' and \code{\link{get_na}}.}
}
\value{
A factor variable, including variable and value labels, respectively
          a data frame with factor variables (including variable and value labels)
          if \code{x} was a data frame.
}
\description{
This function converts a variable into a factor, but keeps
               variable and value labels, if these are attached as attributes
               to the variale. See 'Examples'.
}
\details{
See 'Details' in \code{\link{get_na}}.
}
\note{
This function is intended for use with vectors that have value and variable
       labels attached. Unlike \code{\link{as.factor}}, \code{to_factor} converts
       a variable into a factor and retains the value and variable label attributes.
       \cr \cr
       Attaching labels is automatically done by importing data sets
       with one of the \code{read_*}-functions, like \code{\link{read_spss}}.
       Else, value and variable labels can be manually added to vectors
       with \code{\link{set_labels}} and \code{\link{set_label}}.
}
\examples{
\dontrun{
data(efc)
library(sjPlot)
# normal factor conversion, loses value attributes
efc$e42dep <- as.factor(efc$e42dep)
sjt.frq(efc$e42dep)

data(efc)
# factor conversion, which keeps value attributes
efc$e42dep <- to_factor(efc$e42dep)
sjt.frq(efc$e42dep)}

library(haven)
# create labelled integer, with missing flag
x <- labelled(c(1, 2, 1, 3, 4, 1),
              c(Male = 1, Female = 2, Refused = 3, "N/A" = 4),
              c(FALSE, FALSE, TRUE, TRUE))
# to factor, with missing labels
to_factor(x)
# to factor, missings removed
to_factor(x, drop.na = TRUE)
}
\seealso{
\code{\link{to_value}} to convert a factor into a numeric value and
           \code{\link{to_label}} to convert a value into a factor with labelled
           factor levels.
}

