% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/label_conversion.R
\name{to_na}
\alias{to_na}
\title{Convert missing values of labelled variables into NA}
\usage{
to_na(x)
}
\arguments{
\item{x}{a variable (vector), \code{data.frame} or \code{list} of variables
with attached value labels and missing value codes
(see \code{\link[haven]{labelled}}).}
}
\value{
\code{x}, where each value code of missing values is comverted
           to \code{NA}.
}
\description{
This function converts missing values that are still stored as
               original value code into \code{NA}.
}
\details{
\code{to_na} converts values to \code{NA}, which are defined
           as missing through the \code{is_na}-attribute of a vector
           (see \code{\link[haven]{labelled}}). \code{\link{set_na}},
           by contrast, converts those values to \code{NA} that are
           specified in the function's \code{values} argument; hence,
           \code{\link{set_na}} ignores the \code{is_na}-attribute.
           \cr \cr
           Furthermore, see 'Details' in \code{\link{get_values}},
           \code{\link{get_na}} and \code{\link{to_na}}.
}
\note{
This is a convenient function for \code{set_na(x, get_na(x))}.
}
\examples{
library(haven)

# create labelled factor, with missing flag
x <- labelled(c("M", "M", "F", "X", "N/A"),
              c(Male = "M", Female = "F",
                Refused = "X", "Not applicable" = "N/A"),
              c(FALSE, FALSE, TRUE, TRUE))
x
get_na(x)
to_na(x)


# create labelled integer, with missing flag
x <- labelled(c(1, 2, 1, 3, 4, 1),
             c(Male = 1, Female = 2, Refused = 3, "N/A" = 4),
             c(FALSE, FALSE, TRUE, TRUE))
x
get_na(x)
to_na(x)
}
\seealso{
\code{\link{get_na}} to get value codes of missing values.
}

