% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trim.R
\name{trim}
\alias{trim}
\title{Trim leading and trailing whitespaces from strings}
\usage{
trim(x)
}
\arguments{
\item{x}{Character vector or string. Function is vectorized, i.e. vector
         may have a length greater than 1. See 'Examples'.}
}
\value{
Trimmed \code{x}, i.e. with leading and trailing spaces removed.
}
\description{
Trim leading and trailing whitespaces from strings
}
\examples{
trim("white space at end ")
trim(" white space at start and end ")
trim(c(" string1 ", "   string2", "string 3   "))
}

