% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_write.R
\name{write_stata}
\alias{write_stata}
\title{Write content of data frame to STATA dta-file}
\usage{
write_stata(x, path, enc.to.utf8 = TRUE)
}
\arguments{
\item{x}{data frame that should be saved as STATA-file.}

\item{path}{file path to the STATA dataset.}

\item{enc.to.utf8}{logical, if \code{TRUE}, character encoding of variable and
         value labels will be converted to UTF-8.}
}
\description{
This function saves the content of a data frame to an STATA dta-file.
}
\note{
You don't need to take care whether variables have been imported with
        the \code{\link{read_stata}} function from this package or from \code{haven},
        or if you have imported STATA data and
        created new variables. This function does all necessary data preparation
        to write a properly labelled STATA file.
}
\seealso{
\code{\link{write_spss}}
}

